test: add test case
