+++ src/utils/commonUtil.ts
@@ -1,3 +1,5 @@
+import type { DisplayValueType } from '../BaseSelect';
+
 export function toArray<T>(value: T | T[]): T[] {
   if (Array.isArray(value)) {
     return value;
@@ -13,4 +15,21 @@ export const isBrowserClient = process.env.NODE_ENV !== 'test' && isClient;
 
 export function hasValue(value) {
   return value !== undefined && value !== null;
+}
+
+function isTitleType(title: any) {
+  return ['string', 'number'].includes(typeof title);
+}
+
+export function getTitle(item: DisplayValueType): string {
+  let title: string = undefined;
+  if (item) {
+    if (isTitleType(item.title)) {
+      title = item.title.toString();
+    } else if (isTitleType(item.label)) {
+      title = item.label.toString();
+    }
+  }
+
+  return title;
 }
